web/backend/resources/AI/requirements.txt
python-dotenv==1.0.0
langchain==0.0.137
pinecone-client==2.2.1
openai~=1.6.1